Fixed bugs with handling urine sample data via stdin.
One aspect was the handling of missing steroids, where a sample does not
include data for all 32 of the steroids. Previously, they were not included
in the data file, but this commit changes that, and includes them as NaN in
the file. Reading these in also has to be done differently, as "NaN" is
interpretted as a string (as opposed to a number) by octave's scanf
function.