Modified octave scripts that produced the per patient test inputs and original outputs so that file names have no spaces. Instead of using the patient id (some of which contain spaces) files are named numerically.
Also added file suffixes for input and expected output files.

Currently, one of the tests is reported as failing by the runtests script.
I am suspicious that this may be because the input and expected output for
that test case do not have the same base name. The discrepancy in the
result is only small though; [0, 100] versus [1, 99].