Changes to how VM's are launched and how we get their ip addresses.
I would like to move to using eucalyptus (EC2 type API) to launch the VM,
as that would mean that we would no longer need OpenStack access keys;
just the EC2 keys. This is also good because OpenStack requires a password,
which is more clumsy than EC2, which uses public/private keys. Users need
to remember a password with OpenStack, but not with EC2.

Also changed which packages are installed by setup.sh, since the packages
intalled by pip on Ubuntu did not work. The current setup.sh installs
working packages on Ubuntu 14.10