Added a script to generate a large dataset for testing via random permutation of the existing dataset.
The process is quite simple: shuffle the existing samples randomly and
take the sample on top to go in the permuted set. Thus, the characteristics
of the permuted set will be close to the input set, we won't see wildly
varying steroid levels, as would be the case if we were to randomly
generate samples.