Modified the way that classification tasks are submitted to qsub on the HPC version.
Previously, each individual sample would be run as its own qsub job, but
that will not work very well when it comes to submitting a batch of 50
thousand samples for performance testing. I have changed the system to
use just one qsub instead. All of the samples are kept in a specific
directory, and the classify-edward PBS script will split up the contents of
that directory into blocks of a thousand samples or so, and these will be
processed by job.sh running on one of the nodes allocated to us by PBS.